refactor house into semantic segments
