+++ src/components/features/blog/HouseDemo/HouseFoundation.tsx
@@ -1,0 +1,63 @@
+import { Grid, GridItem } from "@chakra-ui/react";
+
+export function HouseFoundation() {
+  return (
+    <Grid
+      as={"footer"}
+      w="100%"
+      bgColor="gray.600"
+      gap="0.2rem"
+      h="2rem"
+      border={"2px solid"}
+      borderColor={"gray.900"}
+      gridTemplateRows={"repeat(2, 1fr)"}
+      gridTemplateColumns={{
+        base: "repeat(10, 1fr)",
+        md: "repeat(20, 1fr)",
+      }}
+      alignItems="center"
+      justifyContent="center"
+    >
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+      <GridItem h="0.75rem" w="1.5rem" bgColor="gray.400" />
+    </Grid>
+  );
+}
